evaluate destroyed instances during a full destroy
During a full destroy we may need to evaluate instances for providers
that reference them. Since these instance are planned to be removed, we
normally don't want to evaluate them since they would be added to
references of the entire resource. If the entire resource is being
destroyed however, it's more likely that this is a full destroy and the
and we need a special case to allow provider's references to be
evaluated.